address: only run libpostal tests with the build tag provided
